drivers/periph_common: change type of loop counter to int for I2C and SPI initialization
I2C_NUMOF and SPI_NUMOF can be 0 if no configuration is provided, but a PIO emulation for those exists.
This triggers a compiler warning that < 0 is always false for unsigned.

+++ drivers/periph_common/init.c
@@ -73,14 +73,14 @@ void periph_init(void)
     }
     /* initialize configured I2C devices */
 #ifdef MODULE_PERIPH_INIT_I2C
-    for (unsigned i = 0; i < I2C_NUMOF; i++) {
+    for (int i = 0; i < (int)I2C_NUMOF; i++) {
         i2c_init(I2C_DEV(i));
     }
 #endif
 
     /* initialize configured SPI devices */
 #ifdef MODULE_PERIPH_INIT_SPI
-    for (unsigned i = 0; i < SPI_NUMOF; i++) {
+    for (int i = 0; i < (int)SPI_NUMOF; i++) {
         spi_init(SPI_DEV(i));
     }
 #endif
